+++ src/navigation/RootNavigation.js
@@ -1,0 +1,13 @@
+/* eslint-disable prettier/prettier */
+
+// RootNavigation.js
+
+import * as React from 'react';
+
+export const navigationRef = React.createRef();
+
+export function navigate(name, params) {
+  navigationRef.current?.navigate(name, params);
+}
+
+// add other navigation functions that you need and export them
